Modern library approach
1. Flow -> Typescript
2. Add eact-native-builder-bob
3. .npmignore -> files (in package.json)
4. Add peerDependenciesMeta
