+++ .github/actions/install-dependencies/action.yml
@@ -13,6 +13,37 @@ runs:
       with:
         node-version: ${{ inputs.node-version }}
 
+
+    - name: Restore example node_modules from cache
+      uses: actions/cache/restore@v3
+      id: restore_example_node_modules
+      with:
+        key: ${{ runner.os }}-yarn-example-${{ hashFiles('./example/yarn.lock') }}
+        restore-keys: ${{ runner.os }}-yarn-example-
+        path: ${{ github.workspace }}/example/node_modules
+
+
+    - name: Install example node_modules
+      if: steps.restore_example_node_modules.outputs.cache-hit != 'true'
+      shell: bash
+      working-directory: example
+      run: |
+        # Retry 3 times before the steps actually fails
+        (echo "===== Install node_modules Attempt:  1 ====" && yarn install --frozen-lockfile) || \
+        (echo "===== Install node_modules Attempt:  2 ====" && yarn install --frozen-lockfile) || \
+        (echo "===== Install node_modules Attempt:  3 ====" && yarn install --frozen-lockfile) || \
+        (echo "===== Install node_modules Step Failed ====" && exit 1)
+
+
+    - name: Save example node_modules to cache
+      if: steps.restore_example_node_modules.outputs.cache-hit != 'true'
+      uses: actions/cache/save@v3
+      with:
+        key: ${{ steps.restore_example_node_modules.outputs.cache-primary-key }}
+        path: ${{ github.workspace }}/example/node_modules
+
+
+
     - name: Restore node_modules from cache
       uses: actions/cache/restore@v3
       id: restore_node_modules
