+++ package.json
@@ -2,11 +2,17 @@
   "name": "@react-native-community/image-editor",
   "version": "2.3.0",
   "description": "React Native Image Editing native modules for iOS & Android",
-  "main": "lib/ImageEditor.js",
-  "typings": "typings/index.d.ts",
+  "main": "lib/commonjs/index.js",
+  "module": "lib/module/index.js",
+  "types": "lib/typescript/index.d.ts",
+  "react-native": "src/index.ts",
+  "source": "src/index.ts",
   "author": "Dawid Urbaniak <dawidu@onet.pl>",
   "contributors": [],
-  "homepage": "https://github.com/react-native-community/react-native-image-editor#readme",
+  "homepage": "https://github.com/callstack/react-native-image-editor#readme",
+  "bugs": {
+    "url": "https://github.com/callstack/react-native-image-editor/issues"
+  },
   "license": "MIT",
   "keywords": [
     "react-native",
@@ -17,105 +23,84 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/react-native-community/react-native-image-editor.git"
+    "url": "https://github.com/callstack/react-native-image-editor.git"
   },
+  "files": [
+    "src",
+    "lib",
+    "!**/__tests__",
+    "!**/__fixtures__",
+    "!**/__mocks__",
+    "typings",
+    "android",
+    "ios",
+    "react-native-image-editor.podspec",
+    "!android/build",
+    "!ios/build"
+  ],
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test:flow": "flow check",
-    "test:eslint": "eslint \"lib/**/*.js\" \"example/**/*.js\"",
+    "ts": "yarn tsc --noEmit",
+    "lint": "eslint '**/*.{js,ts,tsx}'",
     "release": "release-it",
-    "test:detox:android:test:debug": "detox test -c android.emu.debug",
-    "test:detox:android:test:release": "detox test -c android.emu.release",
-    "test:detox:android:build:debug": "detox build -c android.emu.debug",
-    "test:detox:android:build:release": "detox build -c android.emu.release",
-    "test:detox:android:bundle:release": "mkdir -p .tmp && react-native bundle --max-workers 4 --platform android --dev false --entry-file example/index.js --bundle-output .tmp/android-bundle.js",
-    "test:detox:ios:test:debug": "detox test -c ios.sim.debug",
-    "test:detox:ios:test:release": "detox test -c ios.sim.release",
-    "test:detox:ios:build:debug": "detox build -c ios.sim.debug",
-    "test:detox:ios:build:release": "detox build -c ios.sim.release",
-    "test:detox:ios:bundle:release": "mkdir -p .tmp && react-native bundle --max-workers 4 --platform ios --dev false --entry-file example/index.js --bundle-output .tmp/ios-bundle.js",
-    "test:detox:clean": "rimraf example/android/build && rimraf example/android/app/build && rimraf example/android/.gradle && rimraf example/ios/build"
+    "build": "rm -rf lib/* && bob build && prettier --write ./lib",
+    "prepublishOnly": "yarn run build"
   },
   "peerDependencies": {
-    "react": ">=16.0",
-    "react-native": ">=0.57 <0.65"
+    "react": ">=16.0.0",
+    "react-native": ">=0.57.0"
+  },
+  "peerDependenciesMeta": {
+    "react": {
+      "optional": true
+    },
+    "react-native": {
+      "optional": true
+    }
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@react-native-community/eslint-config": "^0.0.7",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "9.0.0",
-    "babel-jest": "24.1.0",
-    "babel-plugin-module-resolver": "^3.1.3",
-    "detox": "^12.4.9",
-    "eslint": "^6.8.0",
-    "flow-bin": "0.86.0",
-    "jest": "24.1.0",
-    "metro-react-native-babel-preset": "0.51.1",
-    "react": "16.6.3",
-    "react-native": "0.58.4",
-    "react-test-renderer": "16.6.3",
-    "release-it": "^10.3.1"
+    "@react-native/eslint-config": "^0.74.0",
+    "@react-native/typescript-config": "^0.74.0",
+    "@typescript-eslint/parser": "^6.9.1",
+    "eslint": "^8.53.0",
+    "metro-react-native-babel-preset": "0.76.8",
+    "prettier": "^2.8.8",
+    "react": "^18.2.0",
+    "react-native": "^0.72.6",
+    "react-native-builder-bob": "^0.23.1",
+    "react-native-test-app": "^2.5.32",
+    "release-it": "^10.3.1",
+    "typescript": "^5.2.2"
   },
   "eslintConfig": {
-    "extends": "@react-native-community",
-    "env": {
-      "es6": true,
-      "jest": true,
-      "node": true
-    },
-    "globals": {
-      "it": true,
-      "expect": true,
-      "element": true,
-      "describe": true,
-      "by": true,
-      "device": true,
-      "beforeAll": true,
-      "beforeEach": true,
-      "afterAll": true,
-      "jest": true,
-      "jasmine": true
-    }
+    "extends": "@react-native"
   },
-  "jest": {
-    "preset": "react-native",
-    "modulePathIgnorePatterns": [
-      "/e2e/"
-    ]
-  },
-  "detox": {
-    "test-runner": "jest",
-    "runner-config": "example/e2e/config.json",
-    "configurations": {
-      "ios.sim.debug": {
-        "binaryPath": "example/ios/build/Build/Products/Debug-iphonesimulator/ImageEditorExample.app",
-        "build": "export RCT_NO_LAUNCH_PACKAGER=true && xcodebuild -project example/ios/ImageEditorExample.xcodeproj -destination 'platform=iOS Simulator,name=iPhone X' -scheme ImageEditorExample -parallelizeTargets -configuration Debug -derivedDataPath example/ios/build  -UseModernBuildSystem=YES",
-        "type": "ios.simulator",
-        "name": "iPhone X"
-      },
-      "ios.sim.release": {
-        "binaryPath": "example/ios/build/Build/Products/Release-iphonesimulator/ImageEditorExample.app",
-        "build": "export RCT_NO_LAUNCH_PACKAGER=true && xcodebuild -project example/ios/ImageEditorExample.xcodeproj -destination 'platform=iOS Simulator,name=iPhone X' -scheme ImageEditorExample -parallelizeTargets -configuration Release -derivedDataPath example/ios/build  -UseModernBuildSystem=YES",
-        "type": "ios.simulator",
-        "name": "iPhone X"
-      },
-      "android.emu.debug": {
-        "binaryPath": "example/android/app/build/outputs/apk/debug/app-debug.apk",
-        "build": "export RCT_NO_LAUNCH_PACKAGER=true && pushd example/android && ./gradlew assembleDebug assembleAndroidTest -DtestBuildType=debug && popd",
-        "type": "android.emulator",
-        "name": "TestingAVD"
-      },
-      "android.emu.release": {
-        "binaryPath": "example/android/app/build/outputs/apk/release/app-release.apk",
-        "build": "export RCT_NO_LAUNCH_PACKAGER=true && pushd example/android && ./gradlew assembleRelease assembleAndroidTest -DtestBuildType=release && popd",
-        "type": "android.emulator",
-        "name": "TestingAVD"
-      }
-    }
+  "prettier": {
+    "singleQuote": true,
+    "trailingComma": "all",
+    "bracketSpacing": false,
+    "jsxBracketSameLine": true,
+    "parser": "flow"
   },
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/"
-  }
+  },
+  "react-native-builder-bob": {
+    "source": "src",
+    "output": "lib",
+    "targets": [
+      "commonjs",
+      "module",
+      [
+        "typescript",
+        {
+          "project": "tsconfig.build.json"
+        }
+      ]
+    ]
+  },
+  "eslintIgnore": [
+    "node_modules/",
+    "lib/"
+  ]
 }
