+++ tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig",
+  "include": ["src"],
+  "compilerOptions": {
+    "noEmit": false,
+  }
+}
